RT-29: Сделать application импортированной константой, перенести конфигы во view
