+++ src/views/LogIn/LogIn.js
@@ -1,3 +1,30 @@
+import { APPLICATION } from '../../main';
+
+const configForLogin = {
+    inputs: [
+        {
+            placeholder: 'Email',
+            type: 'email',
+        },
+        {
+            placeholder: 'Пароль',
+            type: 'password',
+        }
+    ],
+    networksAuth: [
+        {
+            src: 'img/google.svg',
+        },
+        {
+            src: 'img/facebook.svg',
+        },
+        {
+            src: 'img/vk.svg',
+        }
+    ],
+    buttonName: 'Войти'
+}
+
 /** Class representing a login page view. */
 export class LogInView {
     /**
@@ -6,7 +33,6 @@ export class LogInView {
      * @param {Object} data - Parameters for render signup view.
      */
     constructor({
-                    parent = document.body,
                     data = {},
                 } = {}) {
 
@@ -18,7 +44,7 @@ export class LogInView {
      * Render html login page from pug template to parent.
      */
     render() {
-        const template = puglatizer.LogIn.LogIn(this._data)
-        this._parent.innerHTML = template;
+        const template = puglatizer.LogIn.LogIn(configForLogin)
+        APPLICATION.innerHTML = template;
     }
 }
